+++ .gitignore
@@ -1,2 +1,3 @@
 /target
+/core/src/tests/data
 .idea
