+++ core/src/inherent/mod.rs
@@ -1,0 +1,45 @@
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
+//
+// This file is part of the FUDGE project.
+// FUDGE is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version (see http://www.gnu.org/licenses).
+// Centrifuge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+use sp_inherents::CreateInherentDataProviders;
+use sp_runtime::traits::Block as BlockT;
+use sp_std::marker::PhantomData;
+
+pub use timestamp::Inherent as FudgeInherentTimestamp;
+pub use para_parachain::Inherent as FudgeInherentParaParachain;
+pub use relay_parachain::Inherent as FudgeInherentRelayParachain;
+
+mod timestamp;
+mod para_parachain;
+mod relay_parachain;
+
+pub struct InherentProvider<CIDP, Block, ExtraArgs>
+    where
+        CIDP: CreateInherentDataProviders<Block, ExtraArgs>,
+        Block: BlockT
+{
+    pub cidp: CIDP,
+    pub args: ExtraArgs,
+    _phantom: PhantomData<(Block, ExtraArgs)>
+}
+
+impl<CIDP, Block, ExtraArgs> InherentProvider<CIDP, Block, ExtraArgs>
+    where
+        CIDP: CreateInherentDataProviders<Block, ExtraArgs>,
+        Block: BlockT
+{
+    pub fn new() -> Self {
+        todo!()
+    }
+}
+
+
